+++ gx-frc-19/src/org/gravitechx/frc2019/robot/subsystems/drivesubsystem/DrivePipeline.java
@@ -1,10 +1,9 @@
-package org.gravitechx.frc2019.utils.driveutilities;
+package org.gravitechx.frc2019.robot.subsystems.drivesubsystem;
+
+import org.gravitechx.frc2019.utils.driveutilities.DifferentialDriveSignal;
+import org.gravitechx.frc2019.utils.driveutilities.RotationalDriveSignal;
 
 public class DrivePipeline {
-	public DrivePipeline(){
-		
-	}
-	
 	public DifferentialDriveSignal filter(RotationalDriveSignal rotationalDriveSignal) {
 		rotationalDriveSignal.limitValues();
 		return rotationalDriveSignal.toDifferentialDriveSignal();
